TJBUFSIZE(): Fix int overflow w/ large dimensions
This is the same fix from 24e8225ac2049f2c7280a6e986766f95d5b016cf
applied to the legacy TJBUFSIZE() function in the TurboJPEG 1.0 API.

+++ src/turbojpeg.c
@@ -984,7 +984,8 @@ DLLEXPORT unsigned long TJBUFSIZE(int width, int height)
   /* This allows for rare corner cases in which a JPEG image can actually be
      larger than the uncompressed input (we wouldn't mention it if it hadn't
      happened before.) */
-  retval = PAD(width, 16) * PAD(height, 16) * 6ULL + 2048ULL;
+  retval = (unsigned long long)PAD(width, 16) * PAD(height, 16) * 6ULL +
+           2048ULL;
 #if ULLONG_MAX > ULONG_MAX
   if (retval > (unsigned long long)((unsigned long)-1))
     THROWG("Image is too large", (unsigned long)-1);
